backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-telegram-bot>=20.0
httpx>=0.27.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
